src/imputation_agent/agent.py

from __future__ import annotations
import os, json
from typing import Optional
from dotenv import load_dotenv

def plan_and_run(csv_path: str, out_dir: str = "outputs") -> dict:
    """A tiny 'agent' shim that could be swapped to a real LangChain agent.
    For portability, we keep it minimal here: it profiles, then runs the pipeline.
    """
    # If you want a real LangChain agent, uncomment the imports and code below,
    # and ensure you have an LLM provider setup.
    #
    # from langchain_openai import ChatOpenAI
    # from langchain.tools import Tool
    # from langchain.agents import initialize_agent, AgentType
    # from .tools import tool_profile, tool_run_pipeline
    #
    # llm = ChatOpenAI(model="gpt-4o-mini", temperature=0)
    # tools = [
    #     Tool(name="profile_data", func=tool_profile, description="Profile the CSV and return JSON"),
    #     Tool(name="run_pipeline", func=lambda p: tool_run_pipeline(p, out_dir), description="Run imputation pipeline"),
    # ]
    # agent = initialize_agent(tools, llm, agent=AgentType.ZERO_SHOT_REACT_DESCRIPTION, verbose=True)
    # res = agent.invoke({"input": f"Profile {csv_path} and then run the pipeline, save to {out_dir}."})
    # return {"agent_output": res}
    #
    # Minimal non-LLM fallback:
    from .tools import tool_profile, tool_run_pipeline
    prof = json.loads(tool_profile(csv_path))
    run = json.loads(tool_run_pipeline(csv_path, out_dir))
    return {"profile": prof, "run": run}
