requirements.txt

pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
joblib>=1.3
pydantic>=2.6
typer>=0.12
rich>=13.7
# Optional agent
langchain>=0.2
langchain-openai>=0.1
python-dotenv>=1.0
